+++ test/unit/client.test.ts
@@ -547,6 +547,7 @@ describe('disconnect', () => {
 
   beforeEach(() => {
     client = new Client(getDefaultOptions());
+    client.hardReconnectRequired = false;
     client['autoReconnect'] = true;
   });
 
@@ -575,6 +576,7 @@ describe('disconnect', () => {
     await flushPromises();
 
     expect(isResolved).toBeTruthy();
+    expect(client.hardReconnectRequired).toBeTruthy();
     expect(client['autoReconnect']).toBeFalsy();
     expect(client['cancelConnectionAttempt']).toBeTruthy();
   });
@@ -589,6 +591,7 @@ describe('disconnect', () => {
 
     expect(client.http.stopAllRetries).toHaveBeenCalled();
     expect(isResolved).toBeTruthy();
+    expect(client.hardReconnectRequired).toBeTruthy();
     expect(client['autoReconnect']).toBeFalsy();
     expect(client['cancelConnectionAttempt']).toBeTruthy();
   });
@@ -874,7 +877,12 @@ describe('makeConnectionAttempt', () => {
 
   it('should set up stanzaInstance and emit connected event', async () => {
     expect.assertions(8);
-    const fakeInstance = {};
+    const fakeInstance = {
+      on: jest.fn(),
+      stanzas: {
+        define: jest.fn()
+      }
+    };
     getConnectionSpy.mockResolvedValue(fakeInstance);
     prepareSpy.mockResolvedValue(null);
 
@@ -908,7 +916,7 @@ describe('makeConnectionAttempt', () => {
     const fakeInstance = {
       disconnect: disconnectSpy,
       emit: null,
-      originalEmitter: fakeEmit
+      originalEmitter: fakeEmit,
     };
     const cleanupSpy = jest.spyOn(client as any, 'removeStanzaBoundEventHandlers');
     client['boundStanzaDisconnect'] = async () => {};
@@ -959,7 +967,11 @@ describe('makeConnectionAttempt', () => {
     const fakeInstance = {
       disconnect: disconnectSpy,
       emit: null,
+      on: jest.fn(),
       originalEmitter: fakeEmit,
+      stanzas: {
+        define: jest.fn()
+      },
       subscribeToNode: jest.fn().mockResolvedValue({}),
       pinger: {
         stop: jest.fn()
@@ -1022,12 +1034,37 @@ describe('setupConnectionMonitoring', () => {
     serverMonitorMock.mockClear();
   });
 
+  it('adds a watcher for connection transfers', async () => {
+    const opts = {...getDefaultOptions(), useServerSidePings: false};
+    let client = new Client(opts);
+    const fakeStanza = {
+      on: jest.fn(),
+      stanzas: {
+        define: jest.fn()
+      }
+    };
+
+    client['prepareForConnect'] = jest.fn();
+    client['connectionManager'].getNewStanzaConnection = jest.fn().mockResolvedValue(fakeStanza);
+    client['addInateEventHandlers'] = jest.fn();
+    client['proxyStanzaEvents'] = jest.fn();
+    client['extensions'] = [];
+
+    await client['makeConnectionAttempt']();
+    expect(fakeStanza['connectionTransfer']).toBeTruthy();
+  });
+
   it('uses client-side pings if useServerSidePings is false', async () => {
     const opts = {...getDefaultOptions(), useServerSidePings: false};
     let client = new Client(opts);
 
     client['prepareForConnect'] = jest.fn();
-    client['connectionManager'].getNewStanzaConnection = jest.fn().mockResolvedValue({});
+    client['connectionManager'].getNewStanzaConnection = jest.fn().mockResolvedValue({
+      on: jest.fn(),
+      stanzas: {
+        define: jest.fn()
+      }
+    });
     client['addInateEventHandlers'] = jest.fn();
     client['proxyStanzaEvents'] = jest.fn();
     client['extensions'] = [];
@@ -1045,7 +1082,11 @@ describe('setupConnectionMonitoring', () => {
       throw new Error('pretending server-side pings don\'t work');
     };
     client['connectionManager'].getNewStanzaConnection = jest.fn().mockResolvedValue({
-      subscribeToNode: mockSubscribeToNode
+      on: jest.fn(),
+      subscribeToNode: mockSubscribeToNode,
+      stanzas: {
+        define: jest.fn()
+      }
     });
     client['addInateEventHandlers'] = jest.fn();
     client['proxyStanzaEvents'] = jest.fn();
@@ -1062,7 +1103,11 @@ describe('setupConnectionMonitoring', () => {
     client['prepareForConnect'] = jest.fn();
     const subscribeToNodeSpy = jest.fn().mockResolvedValue({});
     client['connectionManager'].getNewStanzaConnection = jest.fn().mockResolvedValue({
-      subscribeToNode: subscribeToNodeSpy
+      on: jest.fn(),
+      subscribeToNode: subscribeToNodeSpy,
+      stanzas: {
+        define: jest.fn()
+      }
     });
     client['addInateEventHandlers'] = jest.fn();
     client['proxyStanzaEvents'] = jest.fn();